Split git code into 2 modules
